requirements.txt
imap_tools==1.11.1
protobuf==7.34.0
python-dotenv==1.2.2
